Handling division errors
It is also possible to check for division by 0 before attempting
division, however comparison of floating point values may not catch
values that are very close to 0 but not 0. Java produces infinite
or not-a-number results for division by 0 without issuing
exceptions, so it is safer to perform division and then check for
those results.