Input error handling
All input error conditions are handled using exceptions. When
numeric values are entered with errors, not-a-number value is
returned. If such value is detected, A or B is not updated.

This mechanism also handles end-of-input condition. To make output
look consistent, a newline is printed on end of input, so if the
user enters <Ctrl-D>, the next message will be printed on the next
line.